backend/requirements.txt
annotated-types==0.7.0
anthropic==0.44.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.7
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
oauthlib==3.2.2
openai==1.60.0
passlib[bcrypt]==1.7.4
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
pytz
